--- src/Components/Copyright.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Typography, Link } from "@material-ui/core";
-
-export default function Copyright() {
-	return (
-		<Typography variant="body2" color="textSecondary" align="center">
-			{'Copyright © '}
-			<Link color="inherit">
-				Your Website
-			</Link>{' '}
-			{new Date().getFullYear()}
-			{'.'}
-		</Typography>
-	);
-}
